fix: Move document analytics mutation and visitor query hooks; refactor imports
