+++ src/hooks/data/documentAccess/useCreateDocumentAnalyticsMutation.ts
